+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0-alpha.25] - 2022-03-17
+
 ### Added
 - Added ``PybricksHub.race_disconnect()`` method.
 
@@ -224,7 +226,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[Unreleased]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.24..HEAD
+[Unreleased]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.25..HEAD
+[1.0.0-alpha.25]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.24...v1.0.0-alpha.25
 [1.0.0-alpha.24]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.23...v1.0.0-alpha.24
 [1.0.0-alpha.23]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.22...v1.0.0-alpha.23
 [1.0.0-alpha.22]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.21...v1.0.0-alpha.22